feat(management): add YellowDog management foundation (#29)
* feat(management): add refactor foundation

* feat(management): extend profile and console coverage

* ci(release): build management server and netman targets

* fix(ci): restore PR test jobs

Build console assets before CI test workflows so DuskmoonBundler manifests exist on clean runners. Also fix netlink helper clippy failures from the failed PR run.

* fix(management): address PR review feedback

Refs #29 review comment 4910038635.

* ci(e2e): run release smoke checks on PR

* test(netman): stabilize PR checker tests

Refs PR #29 CI Tests jobs 85794056466 and 85793092080.

---------

+++ apps/yellow_dog_netman/native/netlink_helper/src/netlink.rs
@@ -663,7 +663,7 @@ mod tests {
     /// Build a minimal valid netlink message header (16 bytes).
     fn make_nlmsghdr(len: u32, msg_type: u16) -> Vec<u8> {
         let mut buf = vec![0u8; len as usize];
-        buf[0..4].copy_from_slice(&(len as u32).to_ne_bytes());
+        buf[0..4].copy_from_slice(&len.to_ne_bytes());
         buf[4..6].copy_from_slice(&msg_type.to_ne_bytes());
         buf
     }
@@ -1854,6 +1854,7 @@ mod tests {
     // --- Rule with Source and Destination NLAs ---
 
     /// Build an RTM_NEWRULE with source and destination NLAs.
+    #[allow(clippy::too_many_arguments)]
     fn make_rtm_newrule_with_src_dst(
         family: u8,
         dst_len: u8,
